wip: session snapshot 2026-07-18 — workout-detail + fitness-predictor fixes
Preservation snapshot of the working tree (a large, entangled WIP that predates
this branch's HEAD, so isolated buildable slices aren't feasible — this keeps it
safe to carve into clean PRs later). Excludes junk (xlsx, locks, pycache).

This session's completed, verified work included here:
- Workout detail: Phase B three-act redesign; split-routing fix — continuous
  runs with watch-pause water breaks now render real mile splits at MOVING pace
  (not fabricated "reps"/elapsed pace). Verified in-sim on the 13.49mi run.
- Observer prompt (parse-workout-structure.v2): never fabricates interval
  structure from GPS bursts (pause≠recovery, rep>3mi impossible, undeclared→
  continuous); intent_pattern only from the athlete's own words. + cassettes
  004/005. NOT yet deployed to prod.
- Fitness predictor: convert() distance-key fix (was silent no-op); "rep pace ≠
  race pace" prompt + distance-aware pace-segment conversion; regression test.
- Race predictions → ONE number (card + detail + modal); CLAUDE.md hard rule #7
  revised to match.
- reschedule-plan Phase 2 coach guidance + eval cassettes.

Known open (not in scope of this snapshot's correctness):
- revert the "drop interval anchors" change; per-distance PR anchoring for
  accuracy; heat/dew + moving-time in the predictor; deploy the Observer prompt.

+++ RunningLog/RunningLog/Coaching/ModelOfYou/ModelOfYouState.swift
@@ -54,7 +54,11 @@ struct ModelOfYouState: Decodable {
     struct RecoveryRead: Decodable {
         var down_week: Bool?
         var hard_sessions_28d: Int?
-        var avg_days_between_hard: Int?
+        // Backend emits this as an AVERAGE (e.g. 5.7), so it must decode as a
+        // Double. It was typed Int?, and a fractional value like 5.7 threw
+        // "Number 5.7 is not representable in Swift" — which failed the WHOLE
+        // ModelOfYouState decode, blanking the entire surface, not just this row.
+        var avg_days_between_hard: Double?
     }
 
     struct ZonePct: Decodable {
